+++ CHANGELOG.md
@@ -4,5 +4,4 @@
 
 ### Added
 
-- British Library tags: 039, 091, 509, 539, 590, 
-  591, 592
+- British Library tags: 039, 091, 509, 539, 590, 591, 592, 594